backend/requirements.txt
astroid==2.2.5
attrs==19.3.0
Automat==0.8.0
certifi==2019.6.16
cffi==1.13.0
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cssselect==1.1.0
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-rest-swagger==2.0.7
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
joblib==0.13.2
lazy-object-proxy==1.4.2
lxml==4.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.4
numpy==1.17.0
oauthlib==3.1.0
openapi-codec==1.3.2
pandas==0.25.1
parsel==1.5.2
Pillow==6.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
PyJWT==1.7.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.0
Scrapy==1.7.3
service-identity==18.1.0
simplejson==3.16.0
six==1.9.0
sklearn==0.0
SQLAlchemy==1.3.10
sqlparse==0.3.0
Twisted==19.7.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
w3lib==1.21.0
wrapt==1.11.2
zope.interface==4.6.0
